move postprocessing functions to x86 pose class
